Refactor into individual files for maintainability
