--- apps/framework-cli/src/infrastructure/redis.rs
@@ -1,1 +1,0 @@
-pub mod redis_client;
