+++ apps/framework-cli/src/infrastructure/redis/redis_client.rs
@@ -61,19 +61,21 @@
 /// (checking if a lock exists, getting lock information). We only need
 /// to acquire a write lock when we're adding or removing locks.
 ///
-use futures::FutureExt;
 use futures::StreamExt;
-use redis::aio::ConnectionManager;
-use redis::{AsyncCommands, Client, RedisError};
+use redis::{AsyncCommands, Client};
 use serde::{Deserialize, Serialize};
-use std::panic::AssertUnwindSafe;
-use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::atomic::Ordering;
 use std::sync::Arc;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-use tokio::sync::Mutex as TokioMutex;
+use std::time::Duration;
 use tokio::sync::RwLock;
 use tokio::task::JoinHandle;
-use tokio::time;
+
+// Import the modules
+use super::connection::ConnectionManagerWrapper;
+use super::leadership::LeadershipManager;
+use super::messaging::MessagingManager;
+use super::mock::MockRedisClient;
+use super::presence::PresenceManager;
 
 // Type alias for complex callback types
 type MessageCallback = Arc<dyn Fn(String) + Send + Sync>;
@@ -98,7 +100,7 @@ type MessageCallback = Arc<dyn Fn(String) + Send + Sync>;
 ///
 /// ```plaintext
 /// let config = RedisConfig {
-///     url: "redis://redis.example.com:6379".to_string(),
+///     url: "redis://my-redis-server:6379".to_string(),
 ///     key_prefix: "my-service".to_string(),
 /// };
 /// ```
@@ -170,962 +172,6 @@ impl Default for RedisConfig {
     }
 }
 
-/// # Connection Module
-///
-/// This module handles Redis connection management, including connection
-/// establishment, health monitoring, and automatic reconnection. It provides
-/// a thread-safe wrapper around the Redis ConnectionManager to ensure
-/// reliable communication with Redis servers.
-///
-/// ## Key Components
-///
-/// - `ConnectionState`: Enum representing the possible states of a Redis connection.
-///
-/// - `ConnectionManagerWrapper`: A wrapper around Redis ConnectionManager that provides:
-///   - Thread-safe access to Redis connections via Tokio mutexes
-///   - Connection health monitoring via periodic pings
-///   - Automatic reconnection with exponential backoff
-///   - Separate connections for regular commands and pub/sub operations
-///
-/// ## Connection Lifecycle
-///
-/// 1. **Initialization**: Connections are established when creating a new
-///                        `ConnectionManagerWrapper`
-/// 2. **Health Monitoring**: Connections are periodically checked with ping commands
-/// 3. **Failure Detection**: Failed pings trigger the reconnection process
-/// 4. **Reconnection**: Automatic reconnection attempts with exponential backoff
-///
-/// ## Usage Example
-///
-/// ```plaintext
-/// let client = redis::Client::open("redis://127.0.0.1:6379")?;
-/// let connection_wrapper = ConnectionManagerWrapper::new(&client).await?;
-///
-/// Check connection health
-/// if !connection_wrapper.ping().await {
-///     Connection is unhealthy, but reconnection will be attempted automatically
-/// }
-/// ```
-///
-/// ## Thread Safety
-///
-/// The ConnectionManagerWrapper is designed to be thread-safe and can be
-/// safely shared across multiple asynchronous tasks. It uses Tokio mutexes
-/// to ensure that only one task can access the connection at a time,
-/// preventing race conditions.
-pub mod connection {
-    use super::*;
-
-    /// Represents the possible states of a Redis connection.
-    ///
-    /// This enum is used to track and communicate the current state of the
-    /// connection to Redis, which is useful for monitoring and diagnostics.
-    pub enum ConnectionState {
-        /// The connection to Redis is established and operational.
-        Connected,
-        /// The connection to Redis has been lost or closed.
-        Disconnected,
-        /// The system is currently attempting to reestablish a lost connection.
-        Reconnecting,
-    }
-
-    /// A thread-safe wrapper around Redis ConnectionManager.
-    ///
-    /// This struct provides a wrapper around the Redis ConnectionManager that
-    /// ensures thread safety and adds features like connection health
-    /// monitoring and automatic reconnection with exponential backoff.
-    #[derive(Clone)]
-    pub struct ConnectionManagerWrapper {
-        /// The main connection used for regular Redis commands.
-        pub connection: Arc<TokioMutex<ConnectionManager>>,
-        /// A separate connection dedicated to pub/sub operations.
-        pub pub_sub: Arc<TokioMutex<ConnectionManager>>,
-        /// Atomic flag indicating the connection state (true = connected).
-        pub state: Arc<AtomicBool>, // true = connected
-    }
-
-    impl ConnectionManagerWrapper {
-        /// Creates a new ConnectionManagerWrapper with the provided
-        /// Redis client.
-        ///
-        /// This method establishes two separate connections to Redis:
-        /// 1. A main connection for regular Redis commands
-        /// 2. A dedicated connection for pub/sub operations
-        ///
-        /// Both connections are wrapped in Tokio mutexes to ensure thread
-        /// safety. The connection state is initially set to connected (true).
-        ///
-        /// # Parameters
-        ///
-        /// - `client` - A reference to a Redis Client instance used to
-        ///    establish connections
-        ///
-        /// # Returns
-        ///
-        /// - `Result<Self, RedisError>` - A new ConnectionManagerWrapper or a
-        ///   Redis error
-        ///
-        /// # Errors
-        ///
-        /// This method will return an error if it fails to establish either
-        /// connection to Redis.
-        pub async fn new(client: &Client) -> Result<Self, RedisError> {
-            // Create connection manager for normal operations
-            let conn = client.get_connection_manager().await?;
-            let pub_sub_conn = client.get_connection_manager().await?;
-            Ok(ConnectionManagerWrapper {
-                connection: Arc::new(TokioMutex::new(conn)),
-                pub_sub: Arc::new(TokioMutex::new(pub_sub_conn)),
-                state: Arc::new(AtomicBool::new(true)),
-            })
-        }
-
-        /// Checks the health of the Redis connection by sending a PING command.
-        ///
-        /// This method attempts to send a PING command to Redis with a timeout
-        /// of 2 seconds. If the PING succeeds, it returns true. If it fails or
-        /// times out, it updates the connection state to disconnected (false)
-        /// and returns false.
-        ///
-        /// # Returns
-        ///
-        /// - `bool` - true if the connection is healthy, false otherwise
-        pub async fn ping(&self) -> bool {
-            let timeout_future = time::timeout(
-                Duration::from_secs(2),
-                AssertUnwindSafe(async {
-                    redis::cmd("PING")
-                        .query_async::<_, String>(&mut *self.connection.lock().await)
-                        .await
-                })
-                .catch_unwind(),
-            );
-
-            match timeout_future.await {
-                Ok(Ok(Ok(_response))) => true,
-                _ => {
-                    self.state.store(false, Ordering::SeqCst);
-                    false
-                }
-            }
-        }
-
-        /// Attempts to reconnect to Redis with exponential backoff.
-        ///
-        /// This method is called when a connection failure is detected. It
-        /// attempts to reestablish the connection with an exponential backoff
-        /// strategy:
-        ///
-        /// 1. Initial backoff of 5 seconds
-        /// 2. Doubles the backoff time after each failed attempt
-        /// 3. Caps the maximum backoff at 60 seconds
-        ///
-        /// Once a connection is successfully established, it updates the
-        /// connection state to connected (true) and returns.
-        ///
-        /// # Parameters
-        ///
-        /// - `client` - A reference to a Redis Client instance used to
-        ///   establish a new connection
-        /// - `_config` - A reference to the Redis configuration
-        ///   (currently unused)
-        pub async fn attempt_reconnection(&self, client: &Client, _config: &RedisConfig) {
-            let mut backoff = 5;
-            while !self.state.load(Ordering::SeqCst) {
-                time::sleep(Duration::from_secs(backoff)).await;
-                if let Ok(new_conn) = client.get_connection_manager().await {
-                    let mut conn_guard = self.connection.lock().await;
-                    *conn_guard = new_conn;
-                    self.state.store(true, Ordering::SeqCst);
-                    break;
-                } else {
-                    backoff = std::cmp::min(backoff * 2, 60);
-                }
-            }
-        }
-    }
-}
-
-/// # Leadership Module
-///
-/// This module implements distributed leadership election and lock management
-/// using Redis. It provides mechanisms for services to coordinate and elect
-/// a leader among multiple instances, ensuring that critical operations are
-/// performed by only one instance at a time.
-///
-/// ## Key Features
-///
-/// - **Distributed Locking**: Atomic lock acquisition and release using Redis
-///   Lua scripts
-/// - **Leader Election**: Allows multiple service instances to elect a leader
-/// - **Lock Renewal**: Supports extending lock ownership with TTL (Time To Live)
-/// - **Lock Verification**: Methods to check if the current instance owns a
-///   specific lock
-///
-/// ## Lock Lifecycle
-///
-/// 1. **Acquisition**: A service instance attempts to acquire a lock with a TTL
-/// 2. **Verification**: The instance can verify it still owns the lock
-/// 3. **Renewal**: The lock can be renewed to extend its lifetime
-/// 4. **Release**: The lock is explicitly released or expires after its TTL
-///
-/// ## Implementation Details
-///
-/// The module uses Redis Lua scripts to ensure atomicity of lock operations.
-/// Each lock has:
-/// - A unique key in Redis
-/// - An instance ID identifying the lock owner
-/// - A TTL (Time To Live) to prevent deadlocks if an instance fails
-///
-/// ## Usage Example
-///
-/// ```plaintext
-/// Create a leadership manager
-/// let leadership = LeadershipManager::new("instance-123".to_string(),
-/// redis_config);
-///
-/// Attempt to acquire a lock
-/// let lock_acquired = leadership.attempt_lock(connection, "critical-operation", 30).await;
-///
-/// if lock_acquired {
-///     Perform leader-only operations
-///     
-///     Renew the lock to maintain leadership
-///     leadership.renew_lock(connection, "critical-operation", "instance-123", 30).await?;
-///     
-///     Release the lock when done
-///     leadership.release_lock(connection, "critical-operation", "instance-123").await?;
-/// }
-/// ```
-///
-/// ## Thread Safety
-///
-/// The LeadershipManager is designed to be thread-safe and can be safely
-/// shared across multiple asynchronous tasks. The Redis operations are atomic,
-/// ensuring consistent lock state across distributed instances.
-pub mod leadership {
-    use super::*;
-    use redis::Script;
-
-    /// Manager for distributed leadership and lock coordination.
-    ///
-    /// This struct provides methods to acquire, verify, renew, and release
-    /// distributed locks across multiple service instances using Redis.
-    pub struct LeadershipManager {
-        /// Unique identifier for this service instance.
-        pub instance_id: String,
-        /// Redis configuration for connecting to the Redis server.
-        pub redis_config: RedisConfig,
-    }
-
-    impl LeadershipManager {
-        /// Creates a new LeadershipManager instance.
-        ///
-        /// # Parameters
-        ///
-        /// - `instance_id` - A unique identifier for this service instance
-        /// - `config` - Redis configuration for connecting to the Redis server
-        ///
-        /// # Returns
-        ///
-        /// A new LeadershipManager instance configured with the provided
-        /// parameters
-        pub fn new(instance_id: String, config: RedisConfig) -> Self {
-            LeadershipManager {
-                instance_id,
-                redis_config: config,
-            }
-        }
-
-        /// Attempts to acquire a distributed lock.
-        ///
-        /// This method tries to acquire a lock with the following logic:
-        /// - If the lock doesn't exist, it's created with this instance as the owner
-        /// - If the lock exists and is owned by this instance, the TTL is refreshed
-        /// - If the lock exists and is owned by another instance, acquisition fails
-        ///
-        /// The operation is performed atomically using a Lua script.
-        ///
-        /// # Parameters
-        ///
-        /// - `conn` - Redis connection manager
-        /// - `lock_key` - Unique identifier for the lock
-        /// - `ttl` - Time To Live in seconds for the lock
-        ///
-        /// # Returns
-        ///
-        /// - `bool` - true if the lock was acquired or already owned, false
-        ///   otherwise
-        pub async fn attempt_lock(
-            &self,
-            conn: Arc<TokioMutex<ConnectionManager>>,
-            lock_key: &str,
-            ttl: i64,
-        ) -> bool {
-            let script = Script::new(
-                r#"
-                local current = redis.call('GET', KEYS[1])
-                if not current then
-                    redis.call('SET', KEYS[1], ARGV[1])
-                    redis.call('EXPIRE', KEYS[1], ARGV[2])
-                    return 1
-                elseif current == ARGV[1] then
-                    redis.call('EXPIRE', KEYS[1], ARGV[2])
-                    return 1
-                else
-                    return 0
-                end
-                "#,
-            );
-            let instance = self.instance_id.clone();
-            let result = {
-                let mut conn_guard = conn.lock().await;
-                script
-                    .key(lock_key)
-                    .arg(instance)
-                    .arg(ttl)
-                    .invoke_async::<_, i32>(&mut *conn_guard)
-                    .await
-            };
-            matches!(result, Ok(val) if val == 1)
-        }
-
-        /// Renews an existing lock by extending its TTL.
-        ///
-        /// This method extends the TTL of a lock only if it's currently owned
-        /// by the specified instance. The operation is performed atomically
-        /// using a Lua script.
-        ///
-        /// # Parameters
-        ///
-        /// - `conn` - Redis connection manager
-        /// - `lock_key` - Unique identifier for the lock
-        /// - `instance_id` - The instance ID that should own the lock
-        /// - `ttl` - New Time To Live in seconds for the lock
-        ///
-        /// # Returns
-        ///
-        /// - `anyhow::Result<bool>` - Ok(true) if the lock was renewed,
-        ///   Ok(false) if not owned by this instance
-        ///
-        /// # Errors
-        ///
-        /// Returns an error if the Redis operation fails
-        pub async fn renew_lock(
-            &self,
-            conn: Arc<TokioMutex<ConnectionManager>>,
-            lock_key: &str,
-            instance_id: &str,
-            ttl: i64,
-        ) -> anyhow::Result<bool> {
-            let script = redis::Script::new(
-                "if redis.call('GET', KEYS[1]) == ARGV[1] then return redis.call('EXPIRE', KEYS[1], ARGV[2]) else return 0 end",
-            );
-            let result: i32 = script
-                .key(lock_key)
-                .arg(instance_id)
-                .arg(ttl)
-                .invoke_async::<_, i32>(&mut *conn.lock().await)
-                .await?;
-            Ok(result == 1)
-        }
-
-        /// Releases a lock if it's owned by the specified instance.
-        ///
-        /// This method deletes a lock only if it's currently owned by the
-        /// specified instance. The operation is performed atomically using
-        /// a Lua script.
-        ///
-        /// # Parameters
-        ///
-        /// - `conn` - Redis connection manager
-        /// - `lock_key` - Unique identifier for the lock
-        /// - `instance_id` - The instance ID that should own the lock
-        ///
-        /// # Returns
-        ///
-        /// - `anyhow::Result<()>` - Ok(()) if the operation was successful
-        ///
-        /// # Errors
-        ///
-        /// Returns an error if the Redis operation fails
-        pub async fn release_lock(
-            &self,
-            conn: Arc<TokioMutex<ConnectionManager>>,
-            lock_key: &str,
-            instance_id: &str,
-        ) -> anyhow::Result<()> {
-            let script = redis::Script::new(
-                "if redis.call('GET', KEYS[1]) == ARGV[1] then\n return redis.call('DEL', KEYS[1])\nelse\n return 0\nend"
-            );
-            let _result: () = script
-                .key(lock_key)
-                .arg(instance_id)
-                .invoke_async(&mut *conn.lock().await)
-                .await?;
-            Ok(())
-        }
-
-        /// Checks if a lock is currently owned by the specified instance.
-        ///
-        /// This method verifies if a lock exists and is owned by the specified
-        /// instance without modifying the lock. The operation is performed
-        /// atomically using a Lua script.
-        ///
-        /// # Parameters
-        ///
-        /// - `conn` - Redis connection manager
-        /// - `lock_key` - Unique identifier for the lock
-        /// - `instance_id` - The instance ID to check ownership against
-        ///
-        /// # Returns
-        ///
-        /// - `anyhow::Result<bool>` - Ok(true) if the lock is owned by the
-        ///   instance, Ok(false) otherwise
-        ///
-        /// # Errors
-        ///
-        /// Returns an error if the Redis operation fails
-        pub async fn has_lock(
-            &self,
-            conn: Arc<TokioMutex<ConnectionManager>>,
-            lock_key: &str,
-            instance_id: &str,
-        ) -> anyhow::Result<bool> {
-            let script = redis::Script::new(
-                "if redis.call('GET', KEYS[1]) == ARGV[1] then return 1 else return 0 end",
-            );
-            let result: i32 = script
-                .key(lock_key)
-                .arg(instance_id)
-                .invoke_async::<_, i32>(&mut *conn.lock().await)
-                .await?;
-            Ok(result == 1)
-        }
-    }
-}
-
-/// # Presence Module
-///
-/// This module implements service instance presence tracking using Redis.
-/// It allows service instances to register their presence and maintain a
-/// heartbeat, enabling other components to detect active instances and handle
-/// instance failures.
-///
-/// ## Key Features
-///
-/// - **Instance Registration**: Services register their presence in Redis
-/// - **Heartbeat Mechanism**: Periodic updates with TTL to maintain presence
-/// - **Automatic Expiration**: TTL-based keys automatically expire when
-///   instances fail
-/// - **Instance Discovery**: Other services can discover active instances
-///
-/// ## Presence Lifecycle
-///
-/// 1. **Registration**: A service instance registers its presence with a
-///    unique key
-/// 2. **Heartbeat**: The instance periodically updates its presence key with
-///    a new TTL
-/// 3. **Expiration**: If an instance fails, its presence key expires after
-///    the TTL
-/// 4. **Discovery**: Other instances can query Redis to find active instances
-///
-/// ## Implementation Details
-///
-/// The module uses Redis TTL keys to track instance presence:
-/// - Each instance has a unique presence key in Redis
-/// - The key contains a timestamp of the last heartbeat
-/// - The key has a TTL (typically 10 seconds) that must be refreshed
-/// - If an instance fails to update its key, it expires automatically
-///
-/// ## Usage Example
-///
-/// ```plaintext
-/// Create a presence manager
-/// let presence = PresenceManager::new("instance-123".to_string(), "my-service".to_string());
-///
-/// Register and maintain presence
-/// tokio::spawn(async move {
-///     let mut interval = tokio::time::interval(Duration::from_secs(5));
-///     loop {
-///         interval.tick().await;
-///         presence.update_presence(connection).await?;
-///     }
-/// });
-/// ```
-///
-/// ## Thread Safety
-///
-/// The PresenceManager is designed to be thread-safe and can be safely shared
-/// across multiple asynchronous tasks. It is also cloneable, allowing it to be
-/// used in multiple concurrent contexts.
-pub mod presence {
-    use super::*;
-    use anyhow::Context;
-
-    /// Manager for service instance presence tracking.
-    ///
-    /// This struct provides methods to register and maintain the presence
-    /// of a service instance in a distributed system using Redis TTL keys.
-    /// It is cloneable to allow sharing across multiple tasks.
-    #[derive(Clone)]
-    pub struct PresenceManager {
-        /// Unique identifier for this service instance.
-        pub instance_id: String,
-        /// Prefix for Redis keys to prevent collisions.
-        pub key_prefix: String,
-    }
-
-    impl PresenceManager {
-        /// Creates a new PresenceManager instance.
-        ///
-        /// # Parameters
-        ///
-        /// - `instance_id` - A unique identifier for this service instance
-        /// - `key_prefix` - Prefix for Redis keys to prevent collisions
-        ///
-        /// # Returns
-        ///
-        /// A new PresenceManager instance configured with the provided
-        /// parameters
-        pub fn new(instance_id: String, key_prefix: String) -> Self {
-            PresenceManager {
-                instance_id,
-                key_prefix,
-            }
-        }
-
-        /// Generates the Redis key used for tracking this instance's presence.
-        ///
-        /// The key follows the format: `{key_prefix}::{instance_id}::presence`
-        ///
-        /// # Returns
-        ///
-        /// A string containing the formatted presence key
-        fn presence_key(&self) -> String {
-            format!("{}::{}::presence", self.key_prefix, self.instance_id)
-        }
-
-        /// Updates the instance's presence in Redis with a new TTL.
-        ///
-        /// This method:
-        /// 1. Generates the presence key for this instance
-        /// 2. Gets the current timestamp
-        /// 3. Sets the key in Redis with the timestamp as value
-        /// 4. Applies a TTL of 10 seconds to the key
-        ///
-        /// This should be called periodically (typically every few seconds)
-        /// to maintain the instance's presence. If the instance fails to
-        /// update its presence, the key will expire after the TTL.
-        ///
-        /// # Parameters
-        ///
-        /// - `conn` - Redis connection manager
-        ///
-        /// # Returns
-        ///
-        /// - `anyhow::Result<()>` - Ok(()) if the operation was successful
-        ///
-        /// # Errors
-        ///
-        /// Returns an error if:
-        /// - Failed to get the current time
-        /// - Redis operation fails
-        pub async fn update_presence(
-            &self,
-            conn: Arc<TokioMutex<ConnectionManager>>,
-        ) -> anyhow::Result<()> {
-            let key = self.presence_key();
-            let now = SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .context("Failed to get current time")?
-                .as_secs();
-            {
-                let mut conn_guard = conn.lock().await;
-                let _: () = conn_guard.set_ex(&key, now, 10).await?;
-            }
-            Ok(())
-        }
-    }
-}
-
-/// # Messaging Module
-///
-/// This module implements pub/sub messaging between service instances using
-/// Redis. It provides a simple and efficient way for instances to communicate
-/// with each other through broadcast messages or targeted instance-specific
-/// messages.
-///
-/// ## Key Features
-///
-/// - **Publish/Subscribe**: Implements the Redis pub/sub pattern
-/// - **Targeted Messaging**: Send messages to specific service instances
-/// - **Broadcast Messaging**: Send messages to all instances
-/// - **Channel Management**: Handles channel naming and message routing
-///
-/// ## Messaging Patterns
-///
-/// The module supports two primary messaging patterns:
-///
-/// 1. **Direct Messaging**: Messages sent to a specific instance
-///    - Channel format: `{key_prefix}::{instance_id}::msgchannel`
-///    - Only the targeted instance receives the message
-///
-/// 2. **Broadcast Messaging**: Messages sent to all instances
-///    - Channel format: `{key_prefix}::msgchannel`
-///    - All instances subscribed to the channel receive the message
-///
-/// ## Implementation Details
-///
-/// The module uses Redis pub/sub channels to implement messaging:
-/// - Each instance subscribes to its own instance-specific channel
-/// - All instances subscribe to a common broadcast channel
-/// - Messages are published to these channels using the Redis PUBLISH command
-/// - Subscribers receive messages asynchronously through callbacks
-///
-/// ## Usage Example
-///
-/// ```plaintext
-/// Create a messaging manager
-/// let messaging = MessagingManager::new("my-service".to_string());
-///
-/// Send a message to a specific instance
-/// messaging.publish_message(
-///     connection,
-///     "instance-123",
-///     "Hello, specific instance!"
-/// ).await?;
-///
-/// Broadcast a message to all instances
-/// messaging.publish_message(
-///     connection,
-///     "msgchannel",
-///     "Hello, all instances!"
-/// ).await?;
-/// ```
-///
-/// ## Thread Safety
-///
-/// The MessagingManager is designed to be thread-safe and can be safely shared
-/// across multiple asynchronous tasks. The Redis pub/sub operations are handled
-/// by the Redis connection manager, which ensures thread safety.
-pub mod messaging {
-    use super::*;
-
-    /// Manager for pub/sub messaging between service instances.
-    ///
-    /// This struct provides methods to publish messages to specific instances
-    /// or broadcast messages to all instances using Redis pub/sub channels.
-    pub struct MessagingManager {
-        /// Prefix for Redis keys and channels to prevent collisions.
-        pub key_prefix: String,
-    }
-
-    impl MessagingManager {
-        /// Creates a new MessagingManager instance.
-        ///
-        /// # Parameters
-        ///
-        /// - `key_prefix` - Prefix for Redis keys and channels to prevent
-        ///   collisions
-        ///
-        /// # Returns
-        ///
-        /// A new MessagingManager instance configured with the provided key
-        /// prefix
-        pub fn new(key_prefix: String) -> Self {
-            MessagingManager { key_prefix }
-        }
-
-        /// Generates the Redis channel name for a specific target.
-        ///
-        /// This method formats the channel name using the pattern:
-        /// `{key_prefix}::{target}::msgchannel`
-        ///
-        /// The target can be:
-        /// - An instance ID for direct messaging
-        /// - "msgchannel" for broadcast messaging
-        ///
-        /// # Parameters
-        ///
-        /// - `target` - The target identifier (instance ID or "msgchannel")
-        ///
-        /// # Returns
-        ///
-        /// A string containing the formatted channel name
-        pub fn channel_for(&self, target: &str) -> String {
-            format!("{}::{}::msgchannel", self.key_prefix, target)
-        }
-
-        /// Publishes a message to a Redis pub/sub channel.
-        ///
-        /// This method sends a message to the specified target channel.
-        /// The target can be an instance ID for direct messaging or
-        /// a broadcast channel identifier for sending to all instances.
-        ///
-        /// # Parameters
-        ///
-        /// - `conn` - Redis connection manager
-        /// - `target` - The target identifier (instance ID or channel name)
-        /// - `message` - The message content to publish
-        ///
-        /// # Returns
-        ///
-        /// - `anyhow::Result<()>` - Ok(()) if the operation was successful
-        ///
-        /// # Errors
-        ///
-        /// Returns an error if the Redis publish operation fails
-        pub async fn publish_message(
-            &self,
-            conn: Arc<TokioMutex<ConnectionManager>>,
-            target: &str,
-            message: &str,
-        ) -> anyhow::Result<()> {
-            let channel = self.channel_for(target);
-            let mut conn_guard = conn.lock().await;
-            let _: () = conn_guard.publish(&channel, message).await?;
-            Ok(())
-        }
-    }
-}
-
-/// # Mock Module
-///
-/// This module provides a fallback implementation of Redis functionality for
-/// testing and situations where a Redis server is unavailable. It implements
-/// an in-memory version of the Redis operations used by the client, allowing
-/// the application to continue functioning even without a Redis connection.
-///
-/// ## Key Features
-///
-/// - **In-Memory Storage**: Uses a thread-safe HashMap to simulate Redis
-///   storage
-/// - **Fallback Mechanism**: Automatically used when Redis connection fails
-/// - **API Compatibility**: Implements the same interface as the real Redis
-///   client
-/// - **Testing Support**: Enables unit testing without requiring a Redis server
-///
-/// ## Implementation Details
-///
-/// The mock implementation:
-/// - Uses a standard Mutex-protected HashMap for thread-safe storage
-/// - Simulates key-value operations, locks, and queue operations
-/// - Provides simplified versions of Redis commands
-/// - Does not implement all Redis features (only those needed by the client)
-///
-/// ## Usage
-///
-/// The mock client is typically used in two scenarios:
-///
-/// 1. **Automatic Fallback**: When Redis connection fails, the client
-///    automatically switches to the mock implementation:
-///
-///    ```rust
-///    // If Redis connection fails, fallback is used automatically
-///    let redis_client = RedisClient::new("service-name", config).await?;
-///    // Operations will use mock if real Redis is unavailable
-///    redis_client.post_queue_message("message", None).await?;
-///    ```
-///
-/// 2. **Testing**: For unit tests, the mock can be explicitly created:
-///
-///    ```plaintext
-///    Create a mock client for testing
-///    let mock_client = MockRedisClient::new();
-///    mock_client.post_queue_message("queue", "test message").await?;
-///    ```
-///
-/// ## Limitations
-///
-/// The mock implementation has several limitations:
-/// - No persistence (data is lost when the application restarts)
-/// - Limited Redis command support (only implements what's needed)
-/// - No distributed capabilities (only works within a single process)
-/// - Simplified locking behavior (not fully Redis-compatible)
-pub mod mock {
-    use anyhow::Result;
-    use std::collections::HashMap;
-    use std::sync::Arc;
-    use tokio::sync::RwLock;
-
-    /// An in-memory mock implementation of Redis client functionality.
-    ///
-    /// This struct provides a fallback implementation when a Redis server is
-    /// unavailable or for testing purposes. It uses a thread-safe HashMap to
-    /// simulate Redis storage and implements simplified versions of the Redis
-    /// operations used by the client.
-    #[derive(Default, Clone, Debug)]
-    pub struct MockRedisClient {
-        /// Thread-safe in-memory storage that simulates Redis.
-        /// Using RwLock instead of Mutex to allow concurrent reads, which
-        /// improves performance in read-heavy scenarios common in Redis
-        /// usage patterns.
-        pub store: Arc<RwLock<HashMap<String, String>>>,
-    }
-
-    impl MockRedisClient {
-        /// Creates a new MockRedisClient instance with an empty storage.
-        ///
-        /// # Returns
-        ///
-        /// A new MockRedisClient instance with an initialized empty HashMap
-        pub fn new() -> Self {
-            Self {
-                store: Arc::new(RwLock::new(HashMap::new())),
-            }
-        }
-
-        /// Simulates updating the presence of a service instance.
-        ///
-        /// This method stores the presence information in the in-memory
-        /// HashMap, mimicking the behavior of the real Redis presence update
-        /// operation.
-        ///
-        /// # Parameters
-        ///
-        /// - `key` - The presence key for the instance
-        /// - `value` - The timestamp value to store
-        ///
-        /// # Returns
-        ///
-        /// - `Result<()>` - Ok if the operation was successful
-        pub async fn update_presence(&self, key: &str, value: u64) -> Result<()> {
-            let mut store = self.store.write().await;
-            store.insert(key.to_string(), value.to_string());
-            Ok(())
-        }
-
-        /// Simulates attempting to acquire a distributed lock.
-        ///
-        /// This method checks if a lock exists and is owned by the specified
-        /// instance. If the lock doesn't exist, it creates it with the
-        /// specified instance as owner.
-        ///
-        /// # Parameters
-        ///
-        /// - `key` - The lock key
-        /// - `instance_id` - The instance ID attempting to acquire the lock
-        ///
-        /// # Returns
-        ///
-        /// - `bool` - true if the lock was acquired or already owned, false
-        ///   otherwise
-        pub async fn attempt_lock(&self, key: &str, instance_id: &str) -> bool {
-            let mut store = self.store.write().await;
-            if let Some(existing) = store.get(key) {
-                return existing == instance_id;
-            }
-            store.insert(key.to_string(), instance_id.to_string());
-            true
-        }
-
-        /// Simulates renewing a distributed lock.
-        ///
-        /// This method checks if the lock exists and is owned by the specified
-        /// instance. Unlike real Redis, it doesn't extend TTL as the mock
-        /// doesn't implement expiration.
-        ///
-        /// # Parameters
-        ///
-        /// - `key` - The lock key
-        /// - `instance_id` - The instance ID that should own the lock
-        ///
-        /// # Returns
-        ///
-        /// - `bool` - true if the lock exists and is owned by the specified
-        ///   instance
-        pub async fn renew_lock(&self, key: &str, instance_id: &str) -> bool {
-            let store = self.store.read().await;
-            if let Some(current) = store.get(key) {
-                return current == instance_id;
-            }
-            false
-        }
-
-        /// Simulates releasing a distributed lock.
-        ///
-        /// This method removes the lock if it exists and is owned by the
-        /// specified instance.
-        ///
-        /// # Parameters
-        ///
-        /// - `key` - The lock key
-        /// - `instance_id` - The instance ID that should own the lock
-        ///
-        /// # Returns
-        ///
-        /// - `Result<()>` - Ok if the operation was successful
-        pub async fn release_lock(&self, key: &str, instance_id: &str) -> Result<()> {
-            let mut store = self.store.write().await;
-            if let Some(current) = store.get(key) {
-                if current == instance_id {
-                    store.remove(key);
-                }
-            }
-            Ok(())
-        }
-
-        /// Simulates posting a message to a queue.
-        ///
-        /// This method appends a message to the specified queue, creating the
-        /// queue if it doesn't exist. Messages are separated by the '|'
-        /// character.
-        ///
-        /// # Parameters
-        ///
-        /// - `queue` - The queue name
-        /// - `message` - The message to post
-        ///
-        /// # Returns
-        ///
-        /// - `Result<()>` - Ok if the operation was successful
-        pub async fn post_queue_message(&self, queue: &str, message: &str) -> Result<()> {
-            let mut store = self.store.write().await;
-            if let Some(existing) = store.get_mut(queue) {
-                existing.push('|');
-                existing.push_str(message);
-            } else {
-                store.insert(queue.to_string(), message.to_string());
-            }
-            Ok(())
-        }
-
-        /// Retrieves and removes the first message from a queue.
-        ///
-        /// This method gets the first message from the specified queue and
-        /// removes it, simulating the behavior of Redis RPOPLPUSH command.
-        ///
-        /// # Parameters
-        ///
-        /// - `queue` - The queue name
-        ///
-        /// # Returns
-        ///
-        /// - `Result<Option<String>>` - The first message in the queue, or
-        ///   None if empty
-        pub async fn get_queue_message(&self, queue: &str) -> Result<Option<String>> {
-            let mut store = self.store.write().await;
-            if let Some(existing) = store.get_mut(queue) {
-                if let Some(pos) = existing.find('|') {
-                    let msg = existing[..pos].to_string();
-                    *existing = existing[pos + 1..].to_string();
-                    return Ok(Some(msg));
-                }
-
-                // If there's no separator but there is content, return the
-                // entire content
-                if !existing.is_empty() {
-                    let msg = existing.clone();
-                    *existing = String::new();
-                    return Ok(Some(msg));
-                }
-            }
-            Ok(None)
-        }
-    }
-}
-
 /// # Redis Client
 ///
 /// The `RedisClient` is the primary interface for interacting with Redis in
@@ -1189,59 +235,48 @@ pub struct RedisClient {
     pub config: RedisConfig,
     pub service_name: String,
     pub instance_id: String,
-    pub connection_manager: connection::ConnectionManagerWrapper,
-    pub leadership_manager: leadership::LeadershipManager,
-    pub presence_manager: presence::PresenceManager,
-    pub messaging_manager: messaging::MessagingManager,
+    pub connection_manager: ConnectionManagerWrapper,
+    pub leadership_manager: LeadershipManager,
+    pub presence_manager: PresenceManager,
+    pub messaging_manager: MessagingManager,
 
     pub message_callbacks: Arc<RwLock<Vec<MessageCallback>>>,
     pub listener_task: Arc<RwLock<Option<JoinHandle<()>>>>,
     pub presence_task: Arc<RwLock<Option<JoinHandle<()>>>>,
     pub connection_monitor_task: Arc<RwLock<Option<JoinHandle<()>>>>,
-    pub fallback: Option<mock::MockRedisClient>,
+    pub fallback: Option<MockRedisClient>,
 
     pub locks: Arc<RwLock<std::collections::HashMap<String, (String, i64)>>>,
 }
 
 impl RedisClient {
     pub async fn new(service_name: String, config: RedisConfig) -> anyhow::Result<Self> {
-        let client = Client::open(config.url.clone());
-        let (connection_manager, fallback) = match client {
-            Ok(c) => match connection::ConnectionManagerWrapper::new(&c).await {
+        let instance_id = uuid::Uuid::new_v4().to_string();
+
+        // Create Redis client
+        let client_result = Client::open(config.url.clone());
+        let (connection_manager, fallback) = match client_result {
+            Ok(c) => match ConnectionManagerWrapper::new(&c).await {
                 Ok(cm) => (cm, None),
-                Err(e) => {
-                    log::error!("<NewRedisClient> Failed to initialize real Redis connection: {}. Falling back to mock client.", e);
-                    (
-                        connection::ConnectionManagerWrapper {
-                            connection: Arc::new(TokioMutex::new(
-                                c.get_connection_manager().await?,
-                            )),
-                            pub_sub: Arc::new(TokioMutex::new(c.get_connection_manager().await?)),
-                            state: Arc::new(AtomicBool::new(false)),
-                        },
-                        Some(mock::MockRedisClient::new()),
-                    )
+                Err(_) => {
+                    // If we can't connect to Redis, use the mock client
+                    let dummy_client = Client::open("redis://localhost:6379").unwrap();
+                    let cm = ConnectionManagerWrapper::new(&dummy_client).await?;
+                    (cm, Some(MockRedisClient::new()))
                 }
             },
-            Err(e) => {
-                log::error!(
-                    "<NewRedisClient> Failed to open Redis client: {}. Falling back to mock client.",
-                    e
-                );
-                let dummy_client = Client::open("redis://127.0.0.1:1")?;
-                let cm = connection::ConnectionManagerWrapper::new(&dummy_client).await?;
-                (cm, Some(mock::MockRedisClient::new()))
+            Err(_) => {
+                // If we can't even create a client, use the mock client
+                let dummy_client = Client::open("redis://localhost:6379").unwrap();
+                let cm = ConnectionManagerWrapper::new(&dummy_client).await?;
+                (cm, Some(MockRedisClient::new()))
             }
         };
 
-        let instance_id =
-            std::env::var("HOSTNAME").unwrap_or_else(|_| uuid::Uuid::new_v4().to_string());
-
-        let leadership_manager =
-            leadership::LeadershipManager::new(instance_id.clone(), config.clone());
-        let presence_manager =
-            presence::PresenceManager::new(instance_id.clone(), config.key_prefix.clone());
-        let messaging_manager = messaging::MessagingManager::new(config.key_prefix.clone());
+        // Create managers
+        let leadership_manager = LeadershipManager::new(instance_id.clone(), config.clone());
+        let presence_manager = PresenceManager::new(instance_id.clone(), config.key_prefix.clone());
+        let messaging_manager = MessagingManager::new(config.key_prefix.clone());
 
         let client = RedisClient {
             config,
